refactor(core): split P2 modules under LOC budget
